requirements.txt
brotli
google-genai
jupyterlab
jupyter_contrib_nbextensions
keras
matplotlib
numpy
nose2
pandas
pip
scipy
scikit-learn
seaborn
tensorflow
tensorflow-metal
torch
torchvision
